Fix chunk names for negative dimensions
* Update s3save and s3load
* Add tests - clean-up bucket after tests, extend tests with different
  dimension sizes, including negative dimension

+++ py_pkg/tests/common.py
@@ -8,6 +8,8 @@ base_metadata = {'version':   '1',
 def delete_prefix(s3_con, prefix):
     result = s3_con.list_objects_v2(Bucket=bucket_name,
                                     Prefix='/'.join((base_prefix, prefix)))
+    if 'Contents' not in result.keys():
+        return
     objects = [{'Key': e['Key']} for e in result['Contents']]
     s3_con.delete_objects(Bucket=bucket_name,
                           Delete={'Objects': objects})
